+++ HornyGCC/Smpte.h
@@ -1,8 +1,8 @@
 #include <exec/types.h>
 
-LONG Smpte2Ticks(BYTE hh, BYTE mm, BYTE ss, BYTE ff);
+int32 Smpte2Ticks(int8 hh, int8 mm, int8 ss, int8 ff);
 
-BYTE Ticks2ff(LONG ticks);
+int8 Ticks2ff(int32 ticks);
 #define Ticks2ss(ticks) ((ticks/600)%60)
 #define Ticks2mm(ticks) ((ticks/36000)%60)
 #define Ticks2hh(ticks) (ticks/2160000)
